refactor(zalouser): move inbound media helpers to a dedicated module
knip --production flags exports that only tests import. Move the pure
helpers extractInboundMedia and resolveInboundImageContentType into
inbound-media.ts so production code imports them across modules, and
drop the unused testing/__testing re-export.
